Further implementation of the protocol

+++ inc/common/player.h
@@ -8,29 +8,30 @@
 #ifndef PLAYER_H_
 #define PLAYER_H_
 
+#include <cstdint>
 #include "common/position.h"
 
 class Player
 {
 public:
-    int getId() const;
+    uint8_t getId() const;
     Direction getDirection() const;
     Position getPosition() const;
-    int getJoinTime() const;
-    int getKills() const;
+    uint32_t getJoinTime() const;
+    uint32_t getKills() const;
     bool hasPlank() const;
 
-    void setId(int id);
+    void setId(uint8_t id);
     void setDirection(Direction dir);
     void setPosition(Position pos);
 
 protected:
-    int _id;
-	Direction _faceDir;
-	Position _pos;
-	int _joinTime;
-    int _kills;
-	bool _hasPlank;
+    uint8_t _id;
+    Direction _faceDir;
+    Position _pos;
+    uint32_t _joinTime;
+    uint32_t _kills;
+    bool _hasPlank;
 };
 
 #endif /* PLAYER_H_ */
